+++ source/extensions/omni.isaac.debug_draw/nodes/OgnIsaacXPrimRadiusVisualizer.cpp
@@ -39,13 +39,8 @@ public:
     {
         auto& state =
             OgnIsaacXPrimRadiusVisualizerDatabase::sPerInstanceState<OgnIsaacXPrimRadiusVisualizer>(nodeObj, instanceId);
-        omni::renderer::IDebugDraw* debugDrawPtr = carb::getCachedInterface<omni::renderer::IDebugDraw>();
-        if (!debugDrawPtr)
-        {
-            CARB_LOG_ERROR("*** OgnIsaacXPrimRadiusVisualizer failed to acquire debugdraw interface\n");
-        }
         state.mLineDrawing = std::make_shared<omni::isaac::debug_draw::drawing::PrimitiveDrawingHelper>(
-            omni::usd::UsdContext::getContext(), debugDrawPtr,
+            omni::usd::UsdContext::getContext(),
             omni::isaac::debug_draw::drawing::PrimitiveDrawingHelper::RenderingMode::eLines);
     }